+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/batch/BulkExecutor.java
@@ -15,7 +15,10 @@ import com.azure.cosmos.ThrottlingRetryOptions;
 import com.azure.cosmos.TransactionalBatchOperationResult;
 import com.azure.cosmos.TransactionalBatchResponse;
 import com.azure.cosmos.implementation.AsyncDocumentClient;
+import com.azure.cosmos.implementation.ImplementationBridgeHelpers;
+import com.azure.cosmos.implementation.RequestOptions;
 import com.azure.cosmos.implementation.apachecommons.lang.tuple.Pair;
+import com.azure.cosmos.implementation.spark.OperationContextAndListenerTuple;
 import io.netty.handler.codec.http.HttpResponseStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -63,6 +66,7 @@ public final class BulkExecutor<TContext> {
 
     private final CosmosAsyncContainer container;
     private final AsyncDocumentClient docClientWrapper;
+    private final OperationContextAndListenerTuple operationListener;
     private final ThrottlingRetryOptions throttlingRetryOptions;
     private final Flux<CosmosItemOperation> inputOperations;
 
@@ -98,6 +102,9 @@ public final class BulkExecutor<TContext> {
         maxMicroBatchConcurrency = bulkOptions.getMaxMicroBatchConcurrency();
         maxMicroBatchInterval = bulkOptions.getMaxMicroBatchInterval();
         batchContext = bulkOptions.getBatchContext();
+        operationListener = ImplementationBridgeHelpers.CosmosBulkProcessingOptionsHelper
+            .getCosmosBulkProcessingOptionAccessor()
+            .getOperationContext(bulkOptions);
 
         // Initialize sink for handling gone error.
         mainSourceCompleted = new AtomicBoolean(false);
@@ -234,8 +241,7 @@ public final class BulkExecutor<TContext> {
                 return ((ItemBulkOperation<?>) itemOperation).getRetryPolicy().shouldRetry(operationResult).flatMap(
                     result -> {
                         if (result.shouldRetry) {
-                            groupSink.next(itemOperation);
-                            return Mono.empty();
+                            return this.enqueueForRetry(result.backOffTime, groupSink, itemOperation);
                         } else {
                             return Mono.just(BridgeInternal.createCosmosBulkOperationResponse(
                                 itemOperation, cosmosBulkItemResponse, this.batchContext));
@@ -281,16 +287,31 @@ public final class BulkExecutor<TContext> {
         return Mono.just(BridgeInternal.createCosmosBulkOperationResponse(itemOperation, exception, this.batchContext));
     }
 
+    private Mono<CosmosBulkOperationResponse<TContext>> enqueueForRetry(
+        Duration backOffTime,
+        FluxSink<CosmosItemOperation> groupSink,
+        CosmosItemOperation itemOperation) {
+
+        if (backOffTime == null || backOffTime.isZero()) {
+            groupSink.next(itemOperation);
+            return Mono.empty();
+        } else {
+            return Mono
+                .delay(backOffTime)
+                .flatMap((dumm) -> {
+                    groupSink.next(itemOperation);
+                    return Mono.empty();
+                });
+        }
+    }
+
     private Mono<CosmosBulkOperationResponse<TContext>> retryOtherExceptions(
         CosmosItemOperation itemOperation, Exception exception, FluxSink<CosmosItemOperation> groupSink,
         CosmosException cosmosException, ItemBulkOperation<?> itemBulkOperation) {
         return itemBulkOperation.getRetryPolicy().shouldRetry(cosmosException).flatMap(result -> {
             if (result.shouldRetry) {
-
-                groupSink.next(itemOperation);
-                return Mono.empty();
+                return this.enqueueForRetry(result.backOffTime, groupSink, itemBulkOperation);
             } else {
-
                 return Mono.just(BridgeInternal.createCosmosBulkOperationResponse(
                     itemOperation, exception, this.batchContext));
             }
@@ -298,9 +319,11 @@ public final class BulkExecutor<TContext> {
     }
 
     private Mono<TransactionalBatchResponse> executeBatchRequest(PartitionKeyRangeServerBatchRequest serverRequest) {
+        RequestOptions options = new RequestOptions();
+        options.setOperationContextAndListenerTuple(operationListener);
 
         return this.docClientWrapper.executeBatchRequest(
-            BridgeInternal.getLink(this.container), serverRequest, null, false);
+            BridgeInternal.getLink(this.container), serverRequest, options, false);
     }
 
     private void completeAllSinks() {