+++ sdk/datafactory/azure-resourcemanager-datafactory/src/main/java/com/azure/resourcemanager/datafactory/models/DelimitedTextSource.java
@@ -121,6 +121,13 @@ public final class DelimitedTextSource extends CopySource {
         return this;
     }
 
+    /** {@inheritDoc} */
+    @Override
+    public DelimitedTextSource withDisableMetricsCollection(Object disableMetricsCollection) {
+        super.withDisableMetricsCollection(disableMetricsCollection);
+        return this;
+    }
+
     /**
      * Validates the instance.
      *