+++ sdk/storage/azure-storage-blob/CHANGELOG.md
@@ -1,8 +1,12 @@
 # Release History
 
-## 12.12.0-beta.2 (Unreleased)
+## 12.13.0-beta.1 (Unreleased)
+- Added support to get a blob client that uses an encryption scope and customer provided key. 
 - Added support for the 2020-10-02 service version.
 - Added support to specify Parquet Input Serialization when querying a blob.
+
+## 12.12.0 (2021-06-09)
+- GA release
 
 ## 12.12.0-beta.1 (2021-05-13)
 - Added support for the 2020-08-04 service version.