+++ sdk/storage/azure-storage-file-datalake/CHANGELOG.md
@@ -1,8 +1,11 @@
 # Release History
 
-## 12.6.0-beta.2 (Unreleased)
+## 12.7.0-beta.1 (Unreleased)
 - Added support for the 2020-10-02 service version.
 - Added support to specify Parquet Input Serialization when querying a file.
+
+## 12.6.0 (2021-06-09)
+- GA release
 
 ## 12.6.0-beta.1 (2021-05-13)
 - Added support for the 2020-08-04 service version.